Feature: publish with loopback
Problem: in some cases, we want to send a message on a pubsub
topic and receive it on the same node (ex: Aleph message).

Solution: add a `loopback` boolean parameter to `publish`.
When set, the client will forward the message to the P2P service
and also publish it on the subscription exchange.